Move trial to its own file
